requirements.txt
streamlit
pandas
pymupdf
openpyxl
requests
Pillow